Segundo intento

+++ src/main/java/classroom/Asignatura.java
@@ -5,7 +5,7 @@ public class Asignatura {
     public String nombre;
     private String N = "nombre";
     public int codigoInterno;
-    public int codigoExterno;
+    public double codigoExterno;
     public Tipo tipo;
 
     public Asignatura() {
@@ -16,7 +16,7 @@ public class Asignatura {
         this("Sin nombre", codigoInterno, 0,tipo);
     }
 
-    public Asignatura(int codigoExterno,Tipo tipo) {
+    public Asignatura(double codigoExterno,Tipo tipo) {
         this("Sin nombre", 0, codigoExterno, tipo);
     }
 
@@ -24,14 +24,14 @@ public class Asignatura {
         this(nombre, 0, 0, tipo);
     }
 
-    public Asignatura(String nombre, int codigoInterno, int codigoExterno,Tipo tipo) {
+    public Asignatura(String nombre, int codigoInterno, double codigoExterno,Tipo tipo) {
         this.nombre = nombre;
         this.codigoInterno = codigoInterno;
         this.codigoExterno = codigoExterno;
         this.tipo = tipo;
     }
 
-    public void cambiarDatos(int codigoInterno, int codigoExterno, String nombre,Tipo tipo) {
+    public void cambiarDatos(int codigoInterno, double codigoExterno, String nombre,Tipo tipo) {
         this.codigoInterno = codigoInterno;
         this.codigoExterno = codigoExterno;
         this.nombre = nombre;
@@ -42,7 +42,7 @@ public class Asignatura {
         this.codigoInterno = codigoInterno;
     }
 
-    public void cambiarDatos(int codigoExterno) {
+    public void cambiarDatos(double codigoExterno) {
         this.codigoExterno = codigoExterno;
     }
